+++ src/ui/filter-popup-numeric.ts
@@ -6,6 +6,8 @@
 
 import { numericRange } from '../enrichments/filter';
 import type { FilterPredicate } from '../utils/visible-rows';
+import { installPopupChrome, positionPopup } from './popup-chrome';
+export { installPopupChrome, positionPopup };
 
 export interface NumericPopupArgs {
   anchorEl: HTMLElement;
@@ -52,8 +54,8 @@ function buildPopupShell(args: NumericPopupArgs): { popup: HTMLDivElement; input
   popup.setAttribute('role', 'dialog');
   popup.setAttribute('aria-label', 'Numeric filter');
 
-  const min = buildLabelledInput('Min', 'number', stringifyNumber(args.current?.min));
-  const max = buildLabelledInput('Max', 'number', stringifyNumber(args.current?.max));
+  const min = buildLabelledInput('Min', 'number', String(args.current?.min ?? ''));
+  const max = buildLabelledInput('Max', 'number', String(args.current?.max ?? ''));
   const hide = buildLabelledInput('Hide empty cells', 'checkbox', !!args.current?.hideEmpty);
 
   const actionsRow = document.createElement('div');
@@ -70,10 +72,6 @@ function buildPopupShell(args: NumericPopupArgs): { popup: HTMLDivElement; input
     clearBtn,
     applyBtn,
   };
-}
-
-function stringifyNumber(n: number | null | undefined): string {
-  return n === null || n === undefined ? '' : String(n);
 }
 
 function makeButton(label: string): HTMLButtonElement {
@@ -134,58 +132,3 @@ export function openNumericFilterPopup(args: NumericPopupArgs): () => void {
   return dispose;
 }
 
-/* ── Shared popup chrome (focus trap, outside-click, Escape) ────────── */
-
-export function installPopupChrome(
-  popup: HTMLElement,
-  anchorEl: HTMLElement,
-  focusables: readonly HTMLElement[],
-  onClose: () => void
-): () => void {
-  let disposed = false;
-
-  const trap = (ev: KeyboardEvent) => {
-    if (ev.key === 'Escape') {
-      ev.stopPropagation();
-      dispose();
-      return;
-    }
-    if (ev.key !== 'Tab' || focusables.length === 0) return;
-    const idx = focusables.indexOf(document.activeElement as HTMLElement);
-    if (ev.shiftKey && idx <= 0) {
-      ev.preventDefault();
-      focusables[focusables.length - 1].focus();
-    } else if (!ev.shiftKey && idx === focusables.length - 1) {
-      ev.preventDefault();
-      focusables[0].focus();
-    }
-  };
-
-  const outside = (ev: MouseEvent) => {
-    if (!popup.contains(ev.target as Node) && ev.target !== anchorEl) {
-      dispose();
-    }
-  };
-
-  popup.addEventListener('keydown', trap);
-  setTimeout(() => document.addEventListener('mousedown', outside), 0);
-
-  function dispose() {
-    if (disposed) return;
-    disposed = true;
-    popup.removeEventListener('keydown', trap);
-    document.removeEventListener('mousedown', outside);
-    if (popup.parentNode) popup.parentNode.removeChild(popup);
-    try { anchorEl.focus(); } catch { /* ignore */ }
-    onClose();
-  }
-
-  return dispose;
-}
-
-export function positionPopup(popup: HTMLElement, anchor: HTMLElement): void {
-  const rect = anchor.getBoundingClientRect();
-  popup.style.position = 'absolute';
-  popup.style.top = `${rect.bottom + window.scrollY + 2}px`;
-  popup.style.left = `${rect.left + window.scrollX}px`;
-}
